Validate the link descriptor
