+++ coordinator/src/app_desc/validator.rs
@@ -15,41 +15,13 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use super::AppDesc;
-use crate::app_desc::{GlobalName, Namespaced};
 use anyhow::bail;
 
 impl AppDesc {
     pub fn validate(&self) -> anyhow::Result<()> {
-        self.sandboxer_specified()?;
         self.tx_owners_are_valid()?;
-        self.host_imports_are_valid()?;
-        self.module_imports_are_valid()?;
 
         Ok(())
-    }
-
-    fn sandboxer_specified(&self) -> anyhow::Result<()> {
-        if !self.default_sandboxer.is_empty() {
-            return Ok(())
-        }
-
-        let modules_without_sandboxer: Vec<_> = self
-            .modules
-            .iter()
-            .filter_map(|(module, setup)| {
-                if setup.sandboxer.is_empty() {
-                    Some(&**module as &str)
-                } else {
-                    None
-                }
-            })
-            .collect();
-
-        if modules_without_sandboxer.is_empty() {
-            return Ok(())
-        }
-
-        bail!("No sandboxer is specified for modules: {}", modules_without_sandboxer.join(", "))
     }
 
     fn tx_owners_are_valid(&self) -> anyhow::Result<()> {
@@ -77,33 +49,5 @@ impl AppDesc {
                     .join(", ")
             )
         }
-    }
-
-    fn host_imports_are_valid(&self) -> anyhow::Result<()> {
-        self.imports_are_valid("The host", &self.host.imports)
-    }
-
-    fn module_imports_are_valid(&self) -> anyhow::Result<()> {
-        for (module, setup) in self.modules.iter() {
-            self.imports_are_valid(&format!("A module, '{}'", module), &setup.imports)?;
-        }
-        Ok(())
-    }
-
-    fn imports_are_valid(&self, importer: &str, imports: &Namespaced<GlobalName>) -> anyhow::Result<()> {
-        for (_to, from) in imports.iter() {
-            let module = from.module();
-            match self.modules.get(from.module()) {
-                Some(setup) => {
-                    let name = from.name();
-                    if !setup.exports.contains_key(name) {
-                        bail!("{} imports non-existing service '{}' from '{}'", importer, name, module)
-                    }
-                }
-                None => bail!("{} imports from non-existing module: {}", importer, module),
-            }
-        }
-
-        Ok(())
     }
 }
